+++ src/wallet/bitcoin.ts
@@ -1,5 +1,5 @@
-import { Amount, Chain, Network, Pool, TX, WalletDB } from "bcoin";
-import Logger from "blgr";
+import axios, { AxiosRequestConfig, Method } from "axios";
+import { toBitcoin } from "satoshi-bitcoin";
 
 /**
  * Interface defining the Bitcoin wallet functionalities needed by the SDK to execute a swap involving Bitcoin.
@@ -15,152 +15,127 @@ export interface BitcoinWallet {
   sendToAddress(
     address: string,
     satoshis: number,
-    network: string
+    network: Network
   ): Promise<string>;
 
   broadcastTransaction(
     transactionHex: string,
-    network: string
+    network: Network
   ): Promise<string>;
 
   getFee(): string;
 }
 
 /**
- * Dummy Bitcoin wallet based on [bcoin]{@link https://github.com/bcoin-org/bcoin}.
+ * Instance of a bitcoind wallet.
  *
  * This is to be used for demos, examples and dev environment only. No safeguards are applied, no data is written on
  * the disk.
  * This is not to be used for mainnet, instead, implement your own {@link BitcoinWallet}
  */
-export class InMemoryBitcoinWallet implements BitcoinWallet {
+export class BitcoindWallet implements BitcoinWallet {
   public static async newInstance(
-    network: string,
-    peerUri: string,
-    hdKey: string
-  ): Promise<InMemoryBitcoinWallet> {
-    const parsedNetwork = Network.get(network);
+    httpUrl: string,
+    username: string,
+    password: string,
+    walletName: string,
+    privkey?: string
+  ): Promise<BitcoindWallet> {
+    const auth = { username, password };
 
-    const logger = new Logger({
-      level: "warning"
-    });
-    const walletdb = new WalletDB({
-      memory: true,
-      witness: true,
-      logger,
-      network: parsedNetwork
-    });
-    const chain = new Chain({
-      spv: true,
-      logger,
-      network: parsedNetwork
-    });
-    const pool = new Pool({
-      chain,
-      logger
-    });
+    const walletExists = await axios
+      .post(
+        httpUrl,
+        {
+          jsonrpc: "1.0",
+          method: "listwallets"
+        },
+        { auth }
+      )
+      .then(res => res.data.result.includes(walletName));
 
-    await logger.open();
-    await pool.open();
-    await walletdb.open();
-    await chain.open();
-    await pool.connect();
-
-    const wallet = await walletdb.create({
-      logger,
-      network: parsedNetwork,
-      master: hdKey
-    });
-
-    const account = await wallet.getAccount(0);
-
-    for (let i = 0; i < 100; i++) {
-      pool.watchAddress(await account.deriveReceive(i).getAddress());
-      pool.watchAddress(await account.deriveChange(i).getAddress());
+    if (!walletExists) {
+      await axios.post(
+        httpUrl,
+        {
+          jsonrpc: "1.0",
+          method: "createwallet",
+          params: [walletName]
+        },
+        { auth }
+      );
     }
 
-    pool.startSync();
+    const rpcRequestConfig = {
+      url: `${httpUrl}/wallet/${walletName}`,
+      method: "post" as Method,
+      auth
+    };
 
-    pool.on("tx", (tx: any) => {
-      walletdb.addTX(tx);
-    });
+    if (!!privkey) {
+      await axios.request({
+        ...rpcRequestConfig,
+        data: { jsonrpc: "1.0", method: "importprivkey", params: [privkey] }
+      });
+    }
 
-    pool.on("block", (block: any) => {
-      walletdb.addBlock(block);
-      if (block.txs.length > 0) {
-        block.txs.forEach((tx: any) => {
-          walletdb.addTX(tx);
-        });
-      }
-    });
-
-    const netAddr = await pool.hosts.addNode(peerUri);
-    const peer = pool.createOutbound(netAddr);
-    pool.peers.add(peer);
-
-    return new InMemoryBitcoinWallet(
-      parsedNetwork,
-      walletdb,
-      pool,
-      chain,
-      wallet
-    );
+    return new BitcoindWallet(rpcRequestConfig);
   }
 
-  private constructor(
-    public readonly network: any,
-
-    private readonly walletdb: any,
-    private readonly pool: any,
-
-    private readonly chain: any,
-    private readonly wallet: any
-  ) {}
+  private constructor(private rpcRequestConfig: AxiosRequestConfig) {}
 
   public async getBalance(): Promise<number> {
-    const balance = await this.wallet.getBalance();
-    // TODO: Balances stay unconfirmed, try to use bcoin.SPVNode (and set node.http to undefined) see if it catches the confirmations
-    const amount = new Amount(balance.toJSON().unconfirmed, "sat");
-    return amount.toBTC(true);
+    const res = await axios.request({
+      ...this.rpcRequestConfig,
+      data: { jsonrpc: "1.0", method: "getbalance", params: [] }
+    });
+    return res.data.result;
   }
 
   public async getAddress(): Promise<string> {
-    const receiveAddress = await this.wallet.receiveAddress(0);
-    return receiveAddress.toString(this.network);
+    const res = await axios.request({
+      ...this.rpcRequestConfig,
+      data: { jsonrpc: "1.0", method: "getnewaddress", params: [] }
+    });
+
+    return res.data.result;
   }
 
   public async sendToAddress(
     address: string,
     satoshis: number,
-    network: string
+    network: Network
   ): Promise<string> {
-    this.assertNetwork(network);
+    await this.assertNetwork(network);
 
-    const transaction = await this.wallet.send({
-      witness: true,
-      outputs: [
-        {
-          address,
-          value: satoshis
-        }
-      ]
+    const res = await axios.request({
+      ...this.rpcRequestConfig,
+      data: {
+        jsonrpc: "1.0",
+        method: "sendtoaddress",
+        params: [address, toBitcoin(satoshis)]
+      }
     });
-    await this.pool.broadcast(transaction);
 
-    return transaction.txid();
+    return res.data.result;
   }
 
   public async broadcastTransaction(
     transactionHex: string,
-    network: string
+    network: Network
   ): Promise<string> {
-    this.assertNetwork(network);
+    await this.assertNetwork(network);
 
-    const transaction = TX.fromRaw(transactionHex, "hex");
+    const res = await axios.request({
+      ...this.rpcRequestConfig,
+      data: {
+        jsonrpc: "1.0",
+        method: "sendrawtransaction",
+        params: [transactionHex]
+      }
+    });
 
-    await this.pool.broadcast(transaction);
-
-    return transaction.txid();
+    return res.data.result;
   }
 
   public getFee(): string {
@@ -169,27 +144,28 @@ export class InMemoryBitcoinWallet implements BitcoinWallet {
   }
 
   public async close(): Promise<void> {
-    const tasks = [];
-    if (this.pool.opened) {
-      tasks.push(this.pool.close());
-    }
-
-    if (this.chain.opened) {
-      tasks.push(this.chain.close());
-    }
-
-    if (this.walletdb.opened) {
-      tasks.push(this.walletdb.close());
-    }
-
-    await Promise.all(tasks);
+    await axios.request({
+      ...this.rpcRequestConfig,
+      data: {
+        jsonrpc: "1.0",
+        method: "unloadwallet",
+        params: []
+      }
+    });
   }
 
-  private assertNetwork(network: string): void {
-    if (network !== this.network.type) {
-      throw new Error(
-        `This wallet is only connected to the ${this.network.type} network and cannot perform actions on the ${network} network`
+  private async assertNetwork(network: Network): Promise<void> {
+    const res = await axios.request({
+      ...this.rpcRequestConfig,
+      data: { jsonrpc: "1.0", method: "getblockchaininfo", params: [] }
+    });
+
+    if (res.data.result.chain !== network) {
+      return Promise.reject(
+        `This wallet is only connected to the ${network} network and cannot perform actions on the ${network} network`
       );
     }
   }
 }
+
+export type Network = "bitcoin" | "testnet" | "regtest";